#comment implementing   the management of application screen with rol…

+++ projectMonitoringBE/src/main/java/com/pfe/projectMonitoringBE/controllers/KeycloakController.java
@@ -9,6 +9,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import javax.ws.rs.NotFoundException;
 import javax.ws.rs.core.Response;
 
 import org.keycloak.admin.client.Keycloak;
@@ -55,7 +56,7 @@ public class KeycloakController {
 	@PostMapping("/user")
 	public ResponseEntity<String> createUser(@RequestParam String username, @RequestParam String email,
 			@RequestParam String firstname, @RequestParam String lastname, @RequestParam String password,
-			@RequestParam Roles role, @RequestParam String address, @RequestParam Integer telephone) {
+			@RequestParam Roles role, @RequestParam String address, @RequestParam String telephone) {
 		if (StringUtils.isEmpty(username) || StringUtils.isEmpty(password) || StringUtils.isEmpty(email)) {
 			return ResponseEntity.badRequest().body("Empty username or password");
 		}
@@ -77,16 +78,28 @@ public class KeycloakController {
 
 		Response result = keycloak.realm(env.getProperty("keycloak.realm")).users().create(userRepresentation);
 		List<UserRepresentation> userRepresentations = keycloak.realm(env.getProperty("keycloak.realm")).users().list();
+
 		int status = result.getStatus();
+		
 		if (status == 201 || status == 200) {
-			Member member = new Member();
-			for (UserRepresentation user : userRepresentations) {
-				if (user.getEmail() != null && user.getEmail().equalsIgnoreCase(email)) {
-					member.setKeycloakId(user.getId());
-					assignRoles(user.getId(), Arrays.asList(role.name()));
-					break;
+			UsersResource userResources = keycloak.realm(env.getProperty("keycloak.realm")).users();
+			
+			userResources.list().forEach(user -> {
+				if(user.getEmail() != null && user.getEmail().equalsIgnoreCase(userRepresentation.getEmail())== true) {
+					userRepresentation.setId(user.getId());
 				}
-			}
+				
+			} );
+			
+			UserResource userResource = keycloak.realm(env.getProperty("keycloak.realm")).users()
+					.get(userRepresentation.getId());
+
+			Member member = new Member();
+			try {
+			UserRepresentation user = userResource.toRepresentation();
+			member.setKeycloakId(user.getId());
+			assignRoles(user.getId(), Arrays.asList(role.name()));
+			userResource.sendVerifyEmail();
 			member.setEmail(email);
 			member.setName(firstname);
 			member.setLastName(lastname);
@@ -94,7 +107,11 @@ public class KeycloakController {
 			member.setTelephone(telephone);
 			member.setAddress(address);
 			memberService.createOrUpdateMember(member);
+			}catch (NotFoundException e) {
+				System.out.println(e);
+		    }
 		}
+		
 		return new ResponseEntity<>(HttpStatus.valueOf(result.getStatus()));
 	}
 
@@ -132,9 +149,11 @@ public class KeycloakController {
 		return userRepresentations;
 	}
 
-	@PutMapping("/user")
+	@GetMapping("/updateUserInformation")
 	public ResponseEntity<UserRepresentation> updateUserDescriptionAttribute(@RequestParam String username,
-			@RequestParam String firstname, @RequestParam String lastname, @RequestParam String password) {
+			@RequestParam String email, @RequestParam String firstname, @RequestParam String lastname,
+			@RequestParam String password, @RequestParam Roles role, @RequestParam String address,
+			@RequestParam String telephone) {
 		Keycloak keycloak = getKeycloakInstance();
 		Optional<UserRepresentation> user = keycloak.realm(env.getProperty("keycloak.realm")).users().search(username)
 				.stream().filter(u -> u.getUsername().equals(username)).findFirst();
@@ -145,12 +164,25 @@ public class KeycloakController {
 			userRepresentation.setFirstName(firstname);
 			userRepresentation.setLastName(lastname);
 			userRepresentation.setEmail(username);
-			/*
-			 * CredentialRepresentation credentials = new CredentialRepresentation();
-			 * credentials.setType(CredentialRepresentation.PASSWORD);
-			 * credentials.setValue(password); credentials.setTemporary(false);
-			 */
+			List<UserRepresentation> userRepresentations = keycloak.realm(env.getProperty("keycloak.realm")).users()
+					.list();
+
+			Member member = new Member();
+			for (UserRepresentation userInformation : userRepresentations) {
+				if (userInformation.getEmail() != null && userInformation.getEmail().equalsIgnoreCase(email)) {
+					member.setKeycloakId(userInformation.getId());
+					assignRoles(userInformation.getId(), Arrays.asList(role.name()));
+					break;
+				}
+			}
+			member.setEmail(email);
+			member.setName(firstname);
+			member.setLastName(lastname);
+			member.setRole(role);
+			member.setTelephone(telephone);
+			member.setAddress(address);
 			userResource.update(userRepresentation);
+
 			return ResponseEntity.ok().body(userRepresentation);
 		} else {
 			return ResponseEntity.badRequest().build();
@@ -161,11 +193,14 @@ public class KeycloakController {
 	public void deleteUser(@RequestParam String username) {
 		Keycloak keycloak = getKeycloakInstance();
 		UsersResource users = keycloak.realm(env.getProperty("keycloak.realm")).users();
-		users.search(username).stream().forEach(user -> {
 
-			keycloak.realm(env.getProperty("keycloak.realm")).users().delete(user.getId());
-			Member member = memberService.findByKeycloakId(user.getId());
-			memberService.deleteMember(member);
+		users.search(username).stream().forEach(user -> {
+			Response result = keycloak.realm(env.getProperty("keycloak.realm")).users().delete(user.getId());
+			if (result.getStatus() == 201 || result.getStatus() == 200) {
+				Member member = memberService.findByKeycloakId(user.getId());
+				memberService.deleteMember(member);
+			}
+
 		});
 	}
 
@@ -182,15 +217,15 @@ public class KeycloakController {
 	@GetMapping("/roles-by-user")
 	public ResponseEntity<List<RoleRepresentation>> getRolesByUser(@RequestParam String username) {
 		Keycloak keycloak = getKeycloakInstance();
+
 		Optional<UserRepresentation> user = keycloak.realm(env.getProperty("keycloak.realm")).users().search(username)
 				.stream().filter(u -> u.getUsername().equals(username)).findFirst();
 		if (user.isPresent()) {
 			UserRepresentation userRepresentation = user.get();
 			UserResource userResource = keycloak.realm(env.getProperty("keycloak.realm")).users()
 					.get(userRepresentation.getId());
-			ClientRepresentation clientRepresentation = keycloak.realm(env.getProperty("keycloak.realm")).clients()
-					.findByClientId(env.getProperty("keycloak.resource")).get(0);
-			List<RoleRepresentation> roles = userResource.roles().clientLevel(clientRepresentation.getId()).listAll();
+
+			List<RoleRepresentation> roles = userResource.roles().realmLevel().listAll();
 			return ResponseEntity.ok(roles);
 		} else {
 			return ResponseEntity.badRequest().build();
